+++ app/(Pages)/(Categories)/categories/[categoryname]/page.tsx
@@ -1,11 +1,29 @@
-import DetailsPage from "@/app/(Pages)/_PageComponents/DetailsPage";
+// app/category/[categoryname]/page.tsx (Server Component)
 
-const CategoryDetailPage = () => {
+import { Metadata } from "next";
+import CategoryDetailPage from "@/app/(Pages)/_PageComponents/CategoryDetailPage";
+
+interface Params {
+  params: {
+    categoryname: string;
+  };
+}
+
+export async function generateMetadata({ params }: Params): Promise<Metadata> {
+  const { categoryname } = params;
+
+  return {
+    title: `${categoryname.charAt(0).toUpperCase() + categoryname.slice(1)} - Category Details`,
+    description: `Discover more about ${categoryname}. Explore details, insights, and more about this category.`,
+  };
+}
+
+const CategoryPage = ({ params }: Params) => {
   return (
     <main className="overflow-x-hidden">
-      <DetailsPage fetchFrom="category" />;
+      <CategoryDetailPage categoryname={params.categoryname} />
     </main>
   );
 };
 
-export default CategoryDetailPage;
+export default CategoryPage;
