+++ app/(Pages)/(Events)/events/[eventName]/page.tsx
@@ -1,12 +1,28 @@
-import React from "react";
-import EventDetails from "../../_EventComponents/EventDetails";
 
-const EventDetailsPage = () => {
+import { Metadata } from "next";
+import EventDetailsPage from "../../_EventComponents/EventDetailsPage";
+
+interface Params {
+  params: {
+    eventId: string;
+  };
+}
+
+export async function generateMetadata({ params }: Params): Promise<Metadata> {
+  const { eventId } = params;
+
+  return {
+    title: `Event ${eventId} - Details`,
+    description: `Get all the details about Event ${eventId}.`,
+  };
+}
+
+const EventPage = ({ params }: Params) => {
   return (
     <main className="overflow-x-hidden">
-      <EventDetails />;
+      <EventDetailsPage eventId={params.eventId} />
     </main>
   );
 };
 
-export default EventDetailsPage;
+export default EventPage;
